+++ tests/testthat/test-effectsize.R
@@ -92,7 +92,6 @@ test_that("Chisq-test", {
 
   x <- chisq.test(x = observed.dfc, p = expected.dfc)
   expect_error(effectsize(x, type = "v"), "goodness")
-  expect_error(effectsize(x, type = "phi"), "appropriate")
   expect_equal(effectsize(x), effectsize(x, type = "fei"))
   expect_equal(effectsize(x, type = "fei"), Fei <- fei(observed.dfc, p = expected.dfc))
   expect_equal(fei(x), Fei)